+++ tests/no-std/build.rs
@@ -1,11 +1,10 @@
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
-use std::{env, fs, path::PathBuf};
+use std::{env, fs, path::PathBuf, time::SystemTime};
 
 fn main() {
     println!("cargo:rerun-if-changed=build.rs");
     println!("cargo:rustc-check-cfg=cfg(mips,arm_compat)");
-    println!(r#"cargo:rustc-check-cfg=cfg(host_os,values("windows"))"#);
 
     let target_arch = &*env::var("CARGO_CFG_TARGET_ARCH").expect("CARGO_CFG_TARGET_ARCH not set");
 
@@ -19,16 +18,17 @@ fn main() {
     let host = env::var("HOST").expect("HOST not set");
     let profile = env::var("PROFILE").expect("PROFILE not set");
     let out_dir: PathBuf = env::var_os("OUT_DIR").expect("OUT_DIR not set").into();
-    let sep = if host.contains("-windows") {
-        println!(r#"cargo:rustc-cfg=host_os="windows""#);
-        '\\'
-    } else {
-        '/'
-    };
+    let sep = if host.contains("-windows") { '\\' } else { '/' };
+    let duration_since_unix_epoch =
+        SystemTime::now().duration_since(SystemTime::UNIX_EPOCH).unwrap().as_secs();
     fs::write(
-        out_dir.join("expected-bin-path"),
+        out_dir.join("expected"),
         format!(
-            "const EXPECTED_BIN_PATH: &str = r\"{sep}{target}{sep}{profile}{sep}no-std-test\";\n"
+            "\
+            const EXPECTED_BIN_PATH: &str = r\"{sep}{target}{sep}{profile}{sep}no-std-test\";\n\
+            #[cfg(not(mips))]
+            const EXPECTED_DURATION_SINCE_UNIX_EPOCH: Duration = Duration::from_secs({duration_since_unix_epoch});\n\
+            "
         ),
     )
     .unwrap();